requirements.txt
numpy==2.3.0
pypng==0.20220715.0
